pyproject.toml
[project]
name = "capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.1",
    "chromadb>=1.5.5",
    "cohere>=5.18.0",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "langgraph-cli[inmem]>=0.4.18",
    "mysql-connector-python>=9.6.0",
    "ollama>=0.6.1",
    "ollama-llm>=0.1",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2",
    "pypdf>=6.9.1",
    "streamlit>=1.55.0",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "wikipedia>=1.4.0",
]
